\name{exercise_13.17}
\docType{data}
\alias{exercise_13.17}
\title{Clinical Trials Data Set for Exercise 13.17}
\description{Clinical trials data set for Exercise 13.17.}
\usage{data(exercise_13.17)}
\format{A data frame containing 16 observations (rows)
  and the following columns:
  \describe{
  \item{Center}{}
  \item{Treatment}{}
  \item{Much_Better}{}
  \item{Better}{}
  \item{Unchanged.Worse}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
