\name{exercise_5.22}
\docType{data}
\alias{exercise_5.22}
\title{Data on Italian Credit Cards, for Exercise 5.22}
\description{Data on Italian credit cards, for Exercise 5.22.}
\usage{data(exercise_5.22)}
\format{A data frame containing 31 observations (rows)
  and the following columns:
  \describe{
  \item{income}{}
  \item{n}{}
  \item{y}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
