\name{exercise_5.3}
\docType{data}
\alias{exercise_5.3}
\title{Rajon Rondo Assists Data Set for Exercise 5.3}
\description{Rajon Rondo assists data set for Exercise 5.3.}
\usage{data(exercise_5.3)}
\format{A data frame containing 77 observations (rows)
  and the following columns:
  \describe{
  \item{assists}{}
  \item{result}{}
  }
}
\source{\url{https://users.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley & Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
