\name{exercise_6.20}
\docType{data}
\alias{exercise_6.20}
\title{Data on Incontinence Study, for Exercise 6.20}
\description{Data on incontinence study, for Exercise 6.20.}
\usage{data(exercise_6.20)}
\format{A data frame containing 21 observations (rows)
  and the following columns:
  \describe{
  \item{y}{}
  \item{x1}{}
  \item{x2}{}
  \item{x3}{}
  }
}
\source{\url{https://users.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley & Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
