\name{section_13.4.4}
\docType{data}
\alias{section_13.4.4}
\title{Data for Example in Section 13.4.4 on Cluster Sampling}
\description{Data for example in Section 13.4.4 on cluster sampling.}
\usage{data(section_13.4.4)}
\format{A data frame containing 96 observations (rows)
  and the following columns:
  \describe{
  \item{nbhd}{}
  \item{satis_1}{}
  \item{satis_2}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
