\name{table_10.1}
\docType{data}
\alias{table_10.1}
\title{Alcohol, Cigarette, and Marijuana Use, by Gender and Race, as in Table 10.1}
\description{Alcohol, cigarette, and marijuana use, by gender and race, as in Table 10.1.}
\usage{data(table_10.1)}
\format{A data frame containing 32 observations (rows)
  and the following columns:
  \describe{
  \item{a}{}
  \item{c}{}
  \item{m}{}
  \item{r}{}
  \item{g}{}
  \item{count}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
