\name{table_11.5}
\docType{data}
\alias{table_11.5}
\title{Migration Data Set of Table 11.5}
\description{Migration data set of Table 11.5.}
\usage{data(table_11.5)}
\format{A data frame containing 16 observations (rows)
  and the following columns:
  \describe{
  \item{row}{}
  \item{column}{}
  \item{count}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
