\name{table_12.1}
\docType{data}
\alias{table_12.1}
\title{Depression Data Set of Table 12.1 in Case Form}
\description{Depression data set of Table 12.1 in case form.}
\usage{data(table_12.1)}
\format{A data frame containing 1020 observations (rows)
  and the following columns:
  \describe{
  \item{case}{}
  \item{severity}{}
  \item{treat}{}
  \item{time}{}
  \item{outcome}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
