\name{table_13.2}
\docType{data}
\alias{table_13.2}
\title{Presidential Election Poll Data Set of Table 13.2}
\description{Presidential election poll data set of Table 13.2.}
\usage{data(table_13.2)}
\format{A data frame containing 51 observations (rows)
  and the following columns:
  \describe{
  \item{state}{}
  \item{pi}{}
  \item{n}{}
  \item{x}{}
  \item{proportion}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
