\name{table_13.8}
\docType{data}
\alias{table_13.8}
\title{Attitudes Toward Leading Crowd Data Set of Table 13.8}
\description{Attitudes toward leading crowd data set of Table 13.8.}
\usage{data(table_13.8)}
\format{A data frame containing 16 observations (rows)
  and the following columns:
  \describe{
  \item{mem1}{}
  \item{att1}{}
  \item{mem2}{}
  \item{att2}{}
  \item{count}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
