\name{table_15.6}
\docType{data}
\alias{table_15.6}
\alias{exercise_15.10}
\title{Grounds for Divorce Data Set of Table 15.6 for Exercise 15.10}
\description{Grounds for divorce data set of Table 15.6 for Exercise 15.10.}
\usage{data(table_15.6)}
\format{A data frame containing 50 observations (rows)
  and the following columns:
  \describe{
  \item{state}{}
  \item{incompat}{}
  \item{cruelty}{}
  \item{desertn}{}
  \item{non_supp}{}
  \item{alcohol}{}
  \item{felony}{}
  \item{impotenc}{}
  \item{insanity}{}
  \item{separate}{}
  }
}
\source{\url{https://users.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley & Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
