\name{table_4.3}
\docType{data}
\alias{table_4.3}
\title{Horseshoe Crab Data Set of Table 4.3}
\description{Horseshoe crab data set of Table 4.3.}
\usage{data(table_4.3)}
\format{A data frame containing 173 observations (rows)
  and the following columns:
  \describe{
  \item{color}{}
  \item{spine}{}
  \item{width}{}
  \item{satell}{}
  \item{weight}{}
  \item{y}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
