\name{table_4.7}
\docType{data}
\alias{table_4.7}
\title{Teratology Study Data Set of Table 4.7}
\description{Teratology study data set of Table 4.7. (Here y is whethera female crab has a satellite (1=yes, 0=no) and weight is ingrams, rather than kg as in the text.  Also, color has values 1-5 with1=light; there were no crabs of color 1, so in the text, color wasre-coded as color - 1 to give values 1, 2, 3, 4.)}
\usage{data(table_4.7)}
\format{A data frame containing 58 observations (rows)
  and the following columns:
  \describe{
  \item{litter}{}
  \item{group}{}
  \item{n}{}
  \item{y}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
