\name{table_6.15}
\docType{data}
\alias{table_6.15}
\alias{exercise_6.8}
\title{Data on Surgery and Sore Throats in Table 6.15, for Exercise 6.8}
\description{Data on surgery and sore throats in Table 6.15, for Exercise 6.8.}
\usage{data(table_6.15)}
\format{A data frame containing 35 observations (rows)
  and the following columns:
  \describe{
  \item{D}{}
  \item{T}{}
  \item{Y}{}
  }
}
\source{\url{http://www.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley \& Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
