\name{table_7.2}
\docType{data}
\alias{table_7.2}
\title{Full Data Set for Table 7.2 on Endometrial Cancer Grade}
\description{Full data set for Table 7.2 on endometrial cancer grade.}
\usage{data(table_7.2)}
\format{A data frame containing 79 observations (rows)
  and the following columns:
  \describe{
  \item{nv}{}
  \item{pi}{}
  \item{eh}{}
  \item{hg}{}
  }
}
\source{\url{https://users.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley & Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
