\name{table_8.1}
\docType{data}
\alias{table_8.1}
\title{Alligator Food Choice Data Set of Table 8.1}
\description{Alligator food choice data set of Table 8.1.}
\usage{data(table_8.1)}
\format{A data frame containing 80 observations (rows)
  and the following columns:
  \describe{
  \item{lake}{}
  \item{gender}{}
  \item{size}{}
  \item{food}{}
  \item{count}{}
  }
}
\source{\url{https://users.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley & Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
