\name{table_9.3}
\docType{data}
\alias{table_9.3}
\title{High School Student Survey Data Set of Table 9.3}
\description{High school student survey data set of Table 9.3.}
\usage{data(table_9.3)}
\format{A data frame containing 8 observations (rows)
  and the following columns:
  \describe{
  \item{a}{}
  \item{c}{}
  \item{m}{}
  \item{count}{}
  }
}
\source{\url{https://users.stat.ufl.edu/~aa/cda/data.html}}
\references{
  Agresti, A. (2013)
  \emph{Categorical Data Analysis}, Third Edition.
  Hoboken, NJ: John Wiley & Sons.
  ISBN: 978-0-470-46363-5.
}
\keyword{datasets}
